requirements/requirements_doc.txt
Sphinx==5.3.0
ansys-sphinx-theme==0.6.1
sphinx-copybutton==0.5
pyyaml==6.0.0
sphinx-jinja==2.0.2 
tabulate==0.9.0
